move TabItem to /componenets/shared
